+++ DBManager/osx_launchers/osx_add.sh
@@ -2,4 +2,6 @@
 
 DIR=$(dirname "$BASH_SOURCE")/..
 
-"$DIR/dbassistant" add $DIR/tempFile_DBAssistantFromReaper.json
+"$DIR/dbassistant" add "$DIR/tempFile_DBAssistantFromReaper.json"
+
+exit 0