+++ DBManager/osx_launchers/osx_deduplicate.sh
@@ -2,4 +2,6 @@
 
 DIR=$(dirname "$BASH_SOURCE")/..
 
-"$DIR/dbassistant" deduplicate $DIR/tempFile_DBAssistantFromReaper.json
+"$DIR/dbassistant" deduplicate "$DIR/tempFile_DBAssistantFromReaper.json"
+
+exit 0