+++ DBManager/osx_launchers/osx_version.sh
@@ -3,3 +3,5 @@
 DIR=$(dirname "$BASH_SOURCE")/..
 
 "$DIR/dbassistant" -V
+
+exit 0